requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
mysqlclient==2.2.0
python-dotenv==1.0.0
selenium==4.18.1
webdriver-manager==4.0.1
pyautogui==0.9.54
pandas==2.2.1
openpyxl==3.1.2
django-encrypted-model-fields==0.6.1
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.0.0
google-ads-admanager>=0.2.0
python-dateutil>=2.8.2
requests>=2.26.0
beautifulsoup4>=4.12.0
cryptography==41.0.3
django-apscheduler==0.6.2
google-auth-httplib2
xlrd
gunicorn==21.2.0
django-redis==6.0.0 